move tikzjax renderers outside of codeview plugin
